Piles!
Groups of cards that don't make up a full deck are called piles.
std::vector<Card> is a pile

I haven't thought about how to turn that into code so it's more of a
naming convention.

Anything referring to a pile means just one pile.
Anything referring to a deck means all piles in the deck.